requirements.txt
flask
openai
flask-cors
gunicorn
ffmpeg-python

